A Very Special modifier: fermented spider eye. Also known as 'oh no you didn't', or Shadowed's Song. Try it with TNT, in a stack of 64 spider eyes. Then try doubling that, probably in creative (you CAN survive by firing once straight up and running like hell without turning back, if you're lucky)
I don't think this can be practical, but it's a lovely gag, and if you need to dig a hole to bedrock in a big hurry…

Also added Bow as a mid-level speed modifier, not as good as Cake. We could also make it as good as Cake but wearing out.