A bit more spam control; snowballs that pass more than 128 blocks from the point they were launched from will be removed now.
I've also throttled the check to make it cheaper; we don't need to be super prompt here. We check 1/16th of the snowballs each time, so a snowball can exist beyond the limits for 16 * 4 * 1/20th -> 3.2 seconds. In theory.

This seems to effectively mitigate the death vortex spam problem, as it removes these snowballs before they can do much harm.