Boom
I thought it’d be fun to put five sulphur above the respawn snowballs.
Now you have a weak but noisy weapon at all times, and can craft TNT
very early. TNT blows up drops so this is not that big a help, and it
pales in comparison to all the other stuff we have. But it means you
can casually visit Snowball just to go boom for a while.