Basic try at random respawn, but no banlist clearing
So this isn’t that bad at spawning players in a random location, but is
lacking a major part: the idea is that the server can run hardcore, but
that it will clear its banlist upon restarting if the config file says
to do so. It’s mostly succeeding in placing new log-ons in random
locations, but it can’t clear the banlist on restart.