A few tweaks
Jetpack is less high speed for higher level players. Snowmen can move
and attack, and are better at not dying of heatstroke. Glass Bottles
don’t assume the server knows they’re not blocks (bug from a user), and
we check for entity-targeting to make sure they are LivingEntities as
XP orbs were producing an exception.